+++ R/colours.R
@@ -1,11 +1,13 @@
 # Functions for working with NSW and DoE named colours
 
-# Named character vector with every defined NSW colour. See
+# Return a named character vector with every defined NSW colour. See
 # data-raw/colourdata.R to update if needed when the brand changes.
-data("colourdata")
-doestyle_cols <- with(colourdata,
-                      purrr::set_names(hex_value,
-                                       colour_name))
+# This needs to be a function in order to access colourdata internally.
+doestyle_cols <- function() {
+  with(doestyle::colourdata,
+       purrr::set_names(hex_value,
+                        colour_name))
+}
 
 #' Get brand colours by name
 #'
@@ -31,7 +33,7 @@ get_colours <- function(...) {
   }
 
   check_colour_names(colour_names)
-  doestyle_cols[colour_names]
+  doestyle_cols()[colour_names]
 }
 
 #  Aliases for get_colours
@@ -63,8 +65,9 @@ list_doestyle_colours <- function() {
 #' @param colour_names A character vector of colour name(s).
 #'
 #' @return A logical vector.
+#' @noRd
 is_doestyle_colour_name <- function(colour_names) {
-  colour_names %in% names(doestyle_cols)
+  colour_names %in% names(doestyle_cols())
 }
 
 #' Abort if provided with invalid colour names
@@ -73,6 +76,7 @@ is_doestyle_colour_name <- function(colour_names) {
 #' colour names that do not exist in the brand colour list.
 #'
 #' @param colour_names A character vector of colour name(s).
+#' @noRd
 check_colour_names <- function(colour_names) {
   # Does colour_names contain only valid doestyle colour names?
   colours_ok <- is_doestyle_colour_name(colour_names)
@@ -87,4 +91,88 @@ check_colour_names <- function(colour_names) {
       colour names."
     ))
   }
+}
+
+#' Get colour names by hex value
+#'
+#' @param ... Hex values. May be supplied as a character vector of hex values or
+#'   as individual arguments.
+#'
+#' @return `get_colour_names()` returns a character vector of colour names.
+#' @noRd
+get_colour_names <- function(...) {
+  if (is.null(c(...))) {
+    cli::cli_abort(c(
+      "x" = "At least one hex value must be supplied to {.code get_colour_names()}"))
+  }
+
+  hex_values <- std_hex(c(...))
+  check_hex_values(hex_values)
+
+  # Invert hex values and their names
+  colour_names <- purrr::set_names(names(doestyle_cols()), doestyle_cols())
+  unname(colour_names[hex_values])
+  }
+
+#' Get secondary colours for named brand colours
+#'
+#' Secondary colours are on-brand colours that can be used for pattern fills or
+#' other visual elements that need contrast against an on-brand fill colour.
+#' Each NSW Design System colour has a corresponding secondary colour available
+#' via `secondary_colours()`.
+#'
+#' @param ... Colour name(s). May be supplied as a character vector or as
+#'   individual arguments.
+#'
+#' @returns `secondary_colours()` returns the hex value of an appropriate
+#'   secondary colour for each on-brand colour name supplied in its arguments.
+#' @export
+#'
+#' @examples
+#' secondary_colours("red-02", "blue-01", "blue-04")
+secondary_colours <- function(...) {
+  colour_names <- c(...)
+  if (is.null(colour_names)) {
+    cli::cli_abort(c(
+      "x" = "At least one colour name must be supplied to {.code secondary()}"))
+  }
+  check_colour_names(colour_names)
+
+  secondary_lookup <- purrr::set_names(doestyle::colourdata$secondary_colour,
+                                       doestyle::colourdata$colour_name)
+
+  secondary_lookup[colour_names] |>
+    unname() |>
+    get_colours()
+}
+
+#' Get text colours for named brand colours
+#'
+#' Text colours are designed for plotting text over an on-brand fill colour.
+#' Each NSW Design System colour has a corresponding text colour available via
+#' `text_colours()`.
+#'
+#' @param ... Colour name(s). May be supplied as a character vector or as
+#'   individual arguments.
+#'
+#' @returns `secondary_colours()` returns the hex value of an appropriate text
+#'   colour for each on-brand colour name supplied in its arguments.
+#' @export
+#'
+#' @examples
+#' text_colours("red-02", "blue-01", "blue-04")
+text_colours <- function(...) {
+  colour_names <- c(...)
+  if (is.null(colour_names)) {
+    cli::cli_abort(c(
+      "x" = "At least one colour name must be supplied to {.code secondary()}"))
+  }
+  check_colour_names(colour_names)
+
+  text_lookup <- purrr::set_names(doestyle::colourdata$text_colour,
+                                  doestyle::colourdata$colour_name)
+
+  text_lookup[colour_names] |>
+    unname() |>
+    get_colours()
 }