+++ R/doestyle-package.R
@@ -4,5 +4,16 @@
 ## usethis namespace: start
 #' @import ggplot2
 ## usethis namespace: end
+NULL
 
-utils::globalVariables(c(names(colourdata), "doe_palettes"))
+#' Internal palettes methods
+#'
+#' @import palettes
+#' @keywords internal
+#' @name palettes-internal
+NULL
+
+utils::globalVariables(
+  c("doe_palettes", "colour_name", "hex_value", "family", "tone", "nsw_brand",
+    "doe_auxillary", "doe_brand", "doe_primary", "secondary_colour",
+    "text_colour", "red", "green", "blue"))